refactor(proposals): move files
