+++ src/features/view-dao-proposals/components/DaoProposalsTable.module.scss
