+++ src/features/view-dao-proposals/components/DaoProposalsTableCard.module.scss
