+++ src/features/view-dao-proposals/components/DaoProposalsTableRow.module.scss
