+++ src/features/view-dao-proposals/index.ts
@@ -1,1 +1,2 @@
-export * from './DaoProposalsTable';
+export { DaoProposalsTable } from './components/DaoProposalsTable';
+export type { DaoProposalsTableProps } from './components/DaoProposalsTable';
