+++ src/pages/Proposal/components/Body/Body.tsx
@@ -1,6 +1,6 @@
 import styles from './Body.module.scss';
 import { MarkdownViewer, Skeleton } from '@zero-tech/zui/components';
-import { formatProposalBody } from '../../../../features/view-dao-proposals/DaoProposalsTable/lib';
+import { formatProposalBody } from 'features/view-dao-proposals/lib';
 import React from 'react';
 
 export interface BodyProps {
