+++ src/pages/Proposal/components/Vote/VoteModal.tsx
@@ -7,7 +7,7 @@ import {
 	useWeb3,
 	useUserProposalVoteData,
 } from '../../../../lib/hooks';
-import { formatVotingPowerAmount } from '../../../../features/view-dao-proposals/DaoProposalsTable/lib';
+import { formatVotingPowerAmount } from '../../../../features/view-dao-proposals/lib';
 
 import { Modal, Wizard } from '@zero-tech/zui/components';
 import { Approve, Deny } from './VoteButtons';
