+++ src/pages/Proposal/components/Votes/Votes.tsx
@@ -2,7 +2,7 @@ import React, { ReactNode, useMemo } from 'react';
 
 import type { zDAO, Proposal, Vote } from '@zero-tech/zdao-sdk';
 import { useWeb3 } from '../../../../lib/hooks';
-import { formatVotingPowerAmount } from '../../../../features/view-dao-proposals/DaoProposalsTable/lib';
+import { formatVotingPowerAmount } from 'features/view-dao-proposals/lib';
 import { getEtherscanUri } from '../../../../lib/util/network';
 
 import { Skeleton } from '@zero-tech/zui/components';
